Moved logic from header component to own hook and adjusted positioning of header title
